Abstract out distortion effect into a module
The effects modules should all be "children" of the volume controller,
which I think should be called something more like the "axis main
controller". I'll give that some thought.

Each effect gets two data streams (RX & TX), and a switch. If the switch
is high, add the desired effect. If not, simply latch the data. In this
example distortion module the RX is the "incoming data" and the TX is
the "outgoing data".

Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>